% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BackgroundFilter.R
\name{BackgrdFilter}
\alias{BackgrdFilter}
\title{Filter an ExpressionSet for Genes above background Vale}
\usage{
BackgrdFilter(eset, method = "mean")
}
\arguments{
\item{eset}{An ExpressionSet}

\item{method}{can be "mean" which will take the mean expression value of all the controls per array and use the median as the cutoff
Quantile will use the .75 quatile instead of the mean}
}
\description{
Filter an ExpressionSet for Genes above background Vale
}
