% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EnsembleAnnotate.R
\name{EnsemblAnnotate}
\alias{EnsemblAnnotate}
\title{Annotate a ExpressioSet}
\usage{
EnsemblAnnotate(eset, transcriptdb, ensemblonly = T)
}
\arguments{
\item{eset}{An ExpressionSet}
}
\description{
Annotate a ExpressioSet
}
