% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetMainProbes.R
\name{GetMainProbes}
\alias{GetMainProbes}
\title{Annotate a ExpressioSet lifted from Affycoretools}
\usage{
GetMainProbes(input)
}
\arguments{
\item{eset}{An ExpressionSet}
}
\description{
Annotate a ExpressioSet lifted from Affycoretools
}
